requirements.txt
gurobipy>=11.0.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dateutil>=2.9.0
streamlit>=1.36.0